Card Game "Set"
In this project, using MVC format, I created a code for users to play a game of Set.

Description of Game:
This is a card game played by a single player. The game involves a deck of cards. The top of the deck is laid out in a grid format. The player must then choose 3 cards to declare as a set, defined later. If a set is correctly declared, then the chosen cards are removed from the grid and replaced with 3 new cards from the deck.